feat: Rename xAPI management commands to avoid conflicts
